requirements.txt
Django==4.1.4
django-crispy-forms==1.14.0
mysqlclient==2.1.1
django-mysql==4.8.0
django-htmx==1.12.2
coverage==6.5.0